Revert "CANS-946 - Use Component Library / Design System layout (#510)" (#513)
This reverts commit a9c8ac68f21c53c56e17611443c5b91aeb19236a.
