+++ spec/page_objects/assessment_form.rb
@@ -5,7 +5,7 @@ require 'page_objects/sections/breadcrumbs'
 
 class AssessmentGlobal < SitePrism::Section
   element :assessment_page_header, 'h1', text: 'CANS Communimetric Assessment Form'
-  element :save_button, '.header-button i.fa-save'
+  element :save_button, 'div.header-buttons-block i.fa-save'
   element :global_save_success_message_box, 'div.global-alert', text: 'Success! '\
                                             'CANS assessment has been saved'
   element :global_complete_message_box, 'div.global-alert', text: 'This assessment was completed '\
@@ -54,7 +54,7 @@ end
 class AssessmentForm < SitePrism::Page
   section :app_globals, AppGlobals, 'body'
   section :global, AssessmentGlobal, 'body'
-  section :breadcrumbs, Breadcrumbs, '.breadcrumbs'
+  section :breadcrumbs, Breadcrumbs, 'div.breadcrumb-container'
   section :header, AssessmentFormHeader, 'div.assessment-form-header-card'
   section :summary, AssessmentSummary, 'div.assessment-summary-card'
   section :footer, AssessmentFormFooter, 'div.form-footer'
